token
SHA1 for verification

+++ server/api.py
@@ -2,7 +2,7 @@ from django.http import HttpResponse
 from django.views.decorators.csrf import csrf_exempt
 import sqlite3
 import json
-from datetime import datetime, timedelta
+from datetime import datetime
 
 
 @csrf_exempt
@@ -19,6 +19,7 @@ def sync(request):
 	Subject = request.POST.get("Subject","")
 	Country = request.POST.get("Country","")
 	Retention = request.POST.get("Retention","")
+	Token = request.POST.get("Token", None)
 
 	if Retention == "":
 		Retention = 0
@@ -31,7 +32,7 @@ def sync(request):
 		return HttpResponse("Error - invalid username")
 
 	if not Streak.isdigit():
-		return HttpResponse("Error - invalid streak value") 
+		return HttpResponse("Error - invalid streak value")
 
 	if not Cards.isdigit():
 		return HttpResponse("Error - invalid cards value")
@@ -57,17 +58,23 @@ def sync(request):
 	except:
 		return HttpResponse("Error - invalid retention value")
 
-	###
+	### Commit to database ###
 
-	if c.execute("SELECT Username FROM Leaderboard WHERE Username = (?)", (User,)).fetchone(): 
-		c.execute("UPDATE Leaderboard SET Streak = (?), Cards = (?), Time_Spend = (?), Sync_Date = (?), Month = (?), Subject = (?), Country = (?), Retention = (?) WHERE Username = (?) ", (Streak, Cards, Time, Sync_Date, Month, Subject, Country, Retention, User))
-		conn.commit()
-		print("Updated entry: " + str(User))
-	else: 
-		c.execute('INSERT INTO Leaderboard (Username, Streak, Cards , Time_Spend, Sync_Date, Month, Subject, Country, Retention) VALUES(?, ?, ?, ?, ?, ?, ?, ?,?)', (User, Streak, Cards, Time, Sync_Date, Month, Subject, Country, Retention))
+	if c.execute("SELECT Username FROM Leaderboard WHERE Username = (?)", (User,)).fetchone():
+		t = c.execute("SELECT Username, Token FROM Leaderboard WHERE Username = (?)", (User,)).fetchone()
+		if t[1] == Token or t[1] is None or t[1] == "":
+			c.execute("UPDATE Leaderboard SET Streak = (?), Cards = (?), Time_Spend = (?), Sync_Date = (?), Month = (?), Subject = (?), Country = (?), Retention = (?), Token = (?) WHERE Username = (?) ", (Streak, Cards, Time, Sync_Date, Month, Subject, Country, Retention, Token, User))
+			conn.commit()
+			print("Updated entry: " + str(User))
+			return HttpResponse("Done!")
+		else:
+			return HttpResponse("Error - invalid token")
+	else:
+		c.execute('INSERT INTO Leaderboard (Username, Streak, Cards , Time_Spend, Sync_Date, Month, Subject, Country, Retention, Token) VALUES(?, ?, ?, ?, ?, ?, ?, ?,?,?)', (User, Streak, Cards, Time, Sync_Date, Month, Subject, Country, Retention, Token))
 		conn.commit()
 		print("Created new entry: " + str(User))
-	return HttpResponse("Done!")
+		return HttpResponse("Done!")
+
 
 @csrf_exempt
 def all_users(request):
@@ -84,7 +91,7 @@ def all_users(request):
 def get_data(request):
 	conn = sqlite3.connect('/home/ankileaderboard/anki_leaderboard_pythonanywhere/Leaderboard.db')
 	c = conn.cursor()
-	c.execute("SELECT * FROM Leaderboard ORDER BY Cards DESC")
+	c.execute("SELECT Username, Streak, Cards , Time_Spend, Sync_Date, Month, Subject, Country, Retention FROM Leaderboard ORDER BY Cards DESC")
 	return HttpResponse(json.dumps(c.fetchall()))
 
 @csrf_exempt
@@ -92,10 +99,13 @@ def delete(request):
 	conn = sqlite3.connect('/home/ankileaderboard/anki_leaderboard_pythonanywhere/Leaderboard.db')
 	c = conn.cursor()
 	User = request.POST.get("Username", "")
-	try:
-		c.execute("DELETE FROM Leaderboard WHERE Username = (?)", (User,))
-		conn.commit()
-		print("Deleted account: " + str(User))
-		return HttpResponse("Deleted")
-	except:
+	Token = request.POST.get("Token", None)
+
+	if c.execute("SELECT Username FROM Leaderboard WHERE Username = (?)", (User,)).fetchone():
+		t = c.execute("SELECT Username, Token FROM Leaderboard WHERE Username = (?)", (User,)).fetchone()
+		if t[1] == Token or t[1] is None or t[1] == "":
+			c.execute("DELETE FROM Leaderboard WHERE Username = (?)", (User,))
+			conn.commit()
+			print("Deleted account: " + str(User))
+			return HttpResponse("Deleted")
 		return HttpResponse("Failed")
